# Resources/openapi.yaml
openapi: "3.1.0"
info:
  title: container-compose API
  version: "v1"
  description: |
    Container REST API for the container-compose daemon (CHAOS-1351).
    Hand-written spec — see docs/plans/native-api-server.md for architecture decisions.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: "http+unix://{socket}"
    description: Unix domain socket (default)
    variables:
      socket:
        default: "%7E%2F.container-compose%2Fapi.sock"
components:
  schemas:
    APIErrorEnvelope:
      type: object
      required: [error, message, code, requestId]
      properties:
        error:
          type: string
          description: Machine-readable error key (e.g. "not_found")
        message:
          type: string
          description: Human-readable description
        code:
          type: string
          description: Short error code (e.g. "E_404")
        requestId:
          type: string
          description: Correlates with X-Request-Id response header
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIErrorEnvelope"
...
paths:
...
  /projects/{name}:
    get:
      summary: Get project detail (ingested or synthesized)
      description: |
        Returns the parsed compose state when the project was ingested via
        `POST /projects/{name}` (source=ingested). Otherwise synthesizes a
        view from registered containers using the project-name prefix
        convention (source=synthesized). 404 only when neither source
        knows the project.
      operationId: getProject
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                type: object
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      summary: Ingest a compose YAML document
      description: |
        Upload a compose YAML body so the daemon owns the parsed project
        state. Subsequent `up`/`down`/`restart` calls operate against the
        ingested document. Idempotent on byte-identical re-uploads (200);
        differing content under the same project name yields 409.

        Documents containing `include:` directives are rejected — uploaded
        compose is evaluated without filesystem context. CHAOS-1426
        (Decision #12 follow-up).
      operationId: ingestProject
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/yaml:
            schema:
              type: string
              format: binary
              description: Compose YAML document (max 1 MiB)
          application/x-yaml:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Idempotent re-upload of byte-identical content
          content:
            application/json:
              schema:
                type: object
        "201":
          description: New project ingested
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Malformed YAML, includes:-directive present, or empty service map
        "409":
          description: Project name already ingested with different content
        "413":
          description: Body exceeds 1 MiB
        "415":
          description: Content-Type is set but not application/yaml or text/yaml
...
  /projects/{name}/up:
    post:
      summary: Bring up a project
      operationId: projectUp
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Project up
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{name}/down:
    post:
      summary: Take down a project
      operationId: projectDown
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Project down
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{name}/restart:
    post:
      summary: Restart a project
      operationId: projectRestart
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Project restarted
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{name}/build:
    post:
      summary: Build project images (NDJSON streaming)
      operationId: projectBuild
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: NDJSON build stream
          content:
            application/x-ndjson:
              schema:
                type: object
  /projects/{name}/pull:
    post:
      summary: Pull project images (NDJSON streaming)
      operationId: projectPull
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: NDJSON pull stream
          content:
            application/x-ndjson:
              schema:
                type: object
  /projects/{name}/services/{service}/scale:
    post:
      summary: Scale a service
      operationId: projectScaleService
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: service
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [replicas]
              properties:
                replicas:
                  type: integer
      responses:
        "200":
          description: Service scaled
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIErrorEnvelope"
        "404":
          $ref: "#/components/responses/NotFound"
...